+++ src/MAMELoader.cpp
@@ -355,7 +355,7 @@ VirtFile* MAMELoader::LoadRomGroup(MAMERomGroupEntry* entry, const string& forma
 	VirtFile* newVirtFile = new VirtFile(destFile, destFileSize, strstream.str());
 	newVirtFile->DontUseLoaders();
 	newVirtFile->DontUseScanners();
-	newVirtFile->SetProPreRatio(0.80);
+	newVirtFile->SetProPreRatio((float)0.80);
 	return newVirtFile;
 }
